+++ Include/imt3d/CPointsBasedObject.h
@@ -26,8 +26,8 @@ public:
 	// reimplemented (imt3d::IPointsBasedObject)
 	PointFormat GetPointFormat() const override;
 	int GetPointsCount() const override;
-	const void* GetPointData(size_t pointIndex) const override;
-	void* GetPointData(size_t pointIndex) override;
+	const void* GetPointData(int pointIndex) const override;
+	void* GetPointData(int pointIndex) override;
 	virtual void* GetData() override;
 	virtual const void* GetData() const override;
 	virtual int GetPointBytesSize() const override;
@@ -51,16 +51,16 @@ public:
 
 protected:
 	bool Create(PointFormat pointFormat);
-	bool Create(PointFormat pointFormat, size_t pointsCount, const void* dataPtr);
-	bool Append(size_t pointsCount, const void* dataPtr);
+	bool Create(PointFormat pointFormat, int pointsCount, const void* dataPtr);
+	bool Append(int pointsCount, const void* dataPtr);
 	int GetDataSize() const;
 	void EnsureCenterCalculated() const;
 	void EnsureCuboidCalculated() const;
 
 	template <typename PointType> const PointType*
-		TGetPointData(size_t pointIndex, bool validOnly = false) const;
+		TGetPointData(int pointIndex, bool validOnly = false) const;
 	template <typename PointType> PointType*
-		TGetPointData(size_t pointIndex, bool validOnly = false);
+		TGetPointData(int pointIndex, bool validOnly = false);
 	template <typename PointType> bool TIsPointValid(const PointType& pointData) const;
 	template <typename PointType> void TEnsureCenterCalculated() const;
 	template <typename PointType> void TEnsureCuboidCalculated() const;
@@ -72,12 +72,12 @@ protected:
 
 protected:
 	static int GetPointBytesSize(PointFormat pointFormat);
-	static int GetBufferSize(PointFormat pointFormat, size_t pointsCount);
+	static int GetBufferSize(PointFormat pointFormat, int pointsCount);
 
 protected:
 	std::vector<uint8_t> m_data;
 	PointFormat m_pointFormat;
-	size_t m_pointsCount;
+	int m_pointsCount;
 
 	mutable i3d::CVector3d m_cloudCenter;
 	mutable CCuboid m_boundingCuboid;
